backend/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.8
Flask==3.1.1
flask-cors==6.0.0
flatbuffers==23.5.26
frozenlist==1.6.0
gast==0.5.4
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.0
gunicorn==23.0.0
h5py==3.10.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
keras==2.15.0
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.4
ml-dtypes==0.2.0
multidict==6.4.3
numpy==1.26.3
oauthlib==3.2.2
openai==0.28.0
opt-einsum==3.3.0
packaging==23.2
propcache==0.3.1
protobuf==4.23.4
psycopg2-binary==2.9.10
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyJWT==2.10.1
python-dotenv==1.1.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
stripe==12.1.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.35.0
termcolor==2.4.0
tqdm==4.67.1
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.1.3
wrapt==1.14.1
yarl==1.20.0
zipp==3.17.0
